+++ packages/internals/manifest/src/index.ts
@@ -1,2 +1,2 @@
-export type { ApiEntry, PackageManifest } from './types'
+export type { ApiEntry, ApiKind, PackageManifest, SemVer } from './types'
 export { defineManifest } from './define'